I have added a new file named dspec_median_subtracted.py, where I have basically copied the dspec.py . Then I added code so that we can do median subtraction per baseline and channel. Then I also normalise the median subtracted DS by the median of each baseline and channel. This is useful because now we can put a threshold on vmax. Anything higher than 2-3 is a flare/RFI.
To maintain backward compatability, I have added a default parameter to
qlookplot, dspec_external and plt_qlook_image. The name of this parameter
is ds_normalised. By default, it is always set to True. If True, then this
new code takes over. Additionally, since the normalised spectrum is unitless,
if ds_normalised is set to True (which is the case for the flare pipeline),
I do not give the colorbar.