Fix the issue that the output time range for FITS files does not match the input time range
fix the issue in eovsa_flare_calib.py: when specifying a time range, ensure that it retrieves all IDB files necessary to cover the entire time range.
fix the issue in eovsa_flare_pipeline.py: if the provided end time is earlier than the start time of the ms file, adjust the end time to be the end time of the ms file.